server/requirements.txt
Flask
Flask-Cors
psycopg2-binary
bcrypt
python-dotenv
psycopg2
pyjwt